API/requirements.txt

moviepy==2.2.1
openai-whisper==20250625
TikTokApi==7.2.1
yt-dlp==2025.12.8


